server/requirements.txt
Flask==3.0.3
Flask-Cors==4.0.1
gspread==6.1.2
google-auth==2.33.0
